+++ applications/fishing-map/src/features/map/map.hooks.ts
@@ -17,6 +17,7 @@ import {
   useFeatureState,
 } from '@globalfishingwatch/react-hooks/dist/use-map-interaction'
 import GFWAPI from '@globalfishingwatch/api-client'
+import { SublayerCombinationMode } from '@globalfishingwatch/fourwings-aggregate'
 import { ENCOUNTER_EVENTS_SOURCE_ID } from 'features/dataviews/dataviews.utils'
 import { selectLocationType } from 'routes/routes.selectors'
 import { HOME, USER, WORKSPACE, WORKSPACES_LIST } from 'routes/routes'
@@ -27,6 +28,10 @@ import { getDatasetTitleByDataview } from 'features/datasets/datasets.utils'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import { selectHighlightedEvent, setHighlightedEvent } from 'features/timebar/timebar.slice'
 import { setHintDismissed } from 'features/help/hints/hints.slice'
+import {
+  selectShowTimeComparison,
+  selectTimeComparisonValues,
+} from 'features/analysis/analysis.selectors'
 import {
   selectDefaultMapGeneratorsConfig,
   WORKSPACES_POINTS_TYPE,
@@ -61,18 +66,27 @@ export const useGeneratorsConnect = () => {
   const { start, end } = useTimerangeConnect()
   const { viewport } = useViewport()
   const generatorsConfig = useSelector(selectDefaultMapGeneratorsConfig)
+  const showTimeComparison = useSelector(selectShowTimeComparison)
+  const timeComparisonValues = useSelector(selectTimeComparisonValues)
 
   return useMemo(() => {
+    let globalConfig: Generators.GlobalGeneratorConfig = {
+      zoom: viewport.zoom,
+      start,
+      end,
+      token: GFWAPI.getToken(),
+    }
+    if (showTimeComparison && timeComparisonValues) {
+      globalConfig = {
+        ...globalConfig,
+        ...timeComparisonValues,
+      }
+    }
     return {
       generatorsConfig,
-      globalConfig: {
-        zoom: viewport.zoom,
-        start,
-        end,
-        token: GFWAPI.getToken(),
-      },
+      globalConfig,
     }
-  }, [generatorsConfig, viewport.zoom, start, end])
+  }, [generatorsConfig, viewport.zoom, start, end, timeComparisonValues, showTimeComparison])
 }
 
 export const useClickedEventConnect = () => {
@@ -292,6 +306,12 @@ export const parseMapTooltipEvent = (
 ) => {
   if (!event || !event.features) return null
 
+  const baseEvent = {
+    point: event.point,
+    latitude: event.latitude,
+    longitude: event.longitude,
+  }
+
   const clusterFeature = event.features.find(
     (f) => f.generatorType === Generators.Type.TileCluster && parseInt(f.properties.count) > 1
   )
@@ -299,9 +319,7 @@ export const parseMapTooltipEvent = (
   // We don't want to show anything else when hovering a cluster point
   if (clusterFeature) {
     return {
-      point: event.point,
-      latitude: event.latitude,
-      longitude: event.longitude,
+      ...baseEvent,
       features: [
         {
           type: clusterFeature.generatorType,
@@ -312,6 +330,22 @@ export const parseMapTooltipEvent = (
   }
 
   const tooltipEventFeatures: TooltipEventFeature[] = event.features.flatMap((feature) => {
+    const baseFeature = {
+      source: feature.source,
+      sourceLayer: feature.sourceLayer,
+      layerId: feature.layerId as string,
+    }
+
+    if (feature.temporalgrid?.sublayerCombinationMode === SublayerCombinationMode.TimeCompare) {
+      return {
+        ...baseFeature,
+        category: DataviewCategory.Comparison,
+        value: event.features[0]?.value,
+        visible: true,
+        unit: 'hours',
+      } as TooltipEventFeature
+    }
+
     let dataview
     if (feature.generatorId === MERGED_ACTIVITY_ANIMATED_HEATMAP_GENERATOR_ID) {
       const { temporalgrid } = feature
@@ -333,9 +367,7 @@ export const parseMapTooltipEvent = (
       // Not needed to create a dataview just for the workspaces list interaction
       if (feature.generatorId && (feature.generatorId as string).includes(WORKSPACE_GENERATOR_ID)) {
         const tooltipWorkspaceFeature: TooltipEventFeature = {
-          source: feature.source,
-          sourceLayer: feature.sourceLayer,
-          layerId: feature.layerId as string,
+          ...baseFeature,
           type: Generators.Type.GL,
           value: feature.properties.label,
           properties: {},
@@ -383,9 +415,7 @@ export const parseMapTooltipEvent = (
 
   if (!tooltipEventFeatures.length) return null
   return {
-    point: event.point,
-    latitude: event.latitude,
-    longitude: event.longitude,
+    ...baseEvent,
     features: tooltipEventFeatures,
   }
 }