+++ packages/ui-components/src/input-date/InputDate.tsx
@@ -2,6 +2,7 @@ import React, { useRef, forwardRef, useImperativeHandle, memo, Ref } from 'react
 import cx from 'classnames'
 import IconButton from '../icon-button'
 import baseStyles from '../input-text/InputText.module.css'
+import { InputSize } from '../input-text/InputText'
 import styles from './InputDate.module.css'
 
 export type InputDateProps = React.InputHTMLAttributes<HTMLInputElement> & {
@@ -10,22 +11,35 @@ export type InputDateProps = React.InputHTMLAttributes<HTMLInputElement> & {
   max?: string
   min?: string
   step?: number
+  inputSize?: InputSize
   onRemove?: (e: React.MouseEvent) => void
 }
 
 const defaultKey = Date.now().toString()
 
 function InputDate(props: InputDateProps, forwardedRef: Ref<HTMLInputElement>) {
-  const { className, value, label, max, min, step, onRemove, ...rest } = props
+  const {
+    className,
+    value,
+    label,
+    max,
+    min,
+    step,
+    onRemove,
+    inputSize = 'default',
+    ...rest
+  } = props
   const inputRef = useRef<HTMLInputElement>(null)
   useImperativeHandle(forwardedRef, () => inputRef.current as HTMLInputElement)
 
+  const yymmddDate = value?.toString().slice(0, 10)
+
   return (
-    <div className={cx(baseStyles.container, styles.container, className)}>
+    <div className={cx(baseStyles.container, styles.container, styles[inputSize], className)}>
       {label && <label htmlFor={label}>{label}</label>}
       <input
         type="date"
-        value={value}
+        value={yymmddDate}
         key={label || defaultKey}
         className={styles.input}
         ref={inputRef}