+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @globalfishingwatchapp/vessel-history
 
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [f69cd7c51]
+  - @globalfishingwatch/api-types@6.4.0
+  - @globalfishingwatch/layer-composer@10.1.0
+  - @globalfishingwatch/react-hooks@15.2.0
+  - @globalfishingwatch/ui-components@12.5.0
+
 ## 1.0.22
 
 ### Patch Changes