+++ linting/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @globalfishingwatch/linting
 
+## 1.8.0
+
+### Minor Changes
+
+- upgrade to stylelint v14
+
 ## 1.7.0
 
 ### Minor Changes